% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{AnnotateGenes}
\alias{AnnotateGenes}
\title{"AnnotateGenes"}
\usage{
AnnotateGenes(transcript.IDs, trinfo)
}
\arguments{
\item{transcript.IDs}{The transcript IDs to annotate.}

\item{trinfo}{The transcript data frame from which to retrieve the annotation symbol.}
}
\description{
The AnnotateGenes function annotates the TC ID to a HGNC Gene symbol.
}
