% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{AnnotateGeneSymbol}
\alias{AnnotateGeneSymbol}
\title{"AnnotateGeneSymbol"}
\usage{
AnnotateGeneSymbol(symbol.to.annotate)
}
\arguments{
\item{symbol.to.annotate}{Gene symbol to annotate to an ensembl region.}
}
\description{
The AnnotateGeneSymbol function annotates HGNC Gene symbol to an ensemble region.
}
