% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{AnnotationExampleData}
\alias{AnnotationExampleData}
\title{Annotation for the Example Data set}
\format{An object of class \code{"data.frame"}.}
\usage{
data(AnnotationExampleData)
}
\description{
A mapping between junctions and probe sets for the example data set in this package.
}
\examples{
data(AnnotationExampleData)
}
