% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CreateOutput}
\alias{CreateOutput}
\title{"CreateOutput"}
\usage{
CreateOutput(ID, Name, Location = "")
}
\arguments{
\item{ID}{A data frame with a "geneID" column.}

\item{Name}{A name for the returned list.}

\item{Location}{The location where the file should be saved.}
}
\value{
A list with the output of the REIDS_ClusterVersion binded together for all genes.
}
\description{
The CreateOutput functions bind the .RData files returned by the REIDS_ClusterVersion together into one list with an element per gene. The function is advised to be used with the CreateOutput.R and CreateOutput.pbs file in the documentation folder.
}
