% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Example_DAGC}
\alias{Example_DAGC}
\title{DABG p-values}
\format{An object of class \code{"data.frame"}.}
\usage{
data(Example_DAGC)
}
\description{
The DABG p-values for the junctions and probe sets of the example data set in this package.
}
\examples{
data(Example_DAGC)
}
\keyword{datasets}
