% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_probeset_probesfile}
\alias{Example_probeset_probesfile}
\title{Probe ID's for a set of probe sets}
\format{An object of class \code{"data.frame"}.}
\usage{
data(Example_probeset_probesfile)
}
\description{
A mapping between the probe ID's and the unit ID's for the DABG file..
}
\examples{
data(Example_probeset_probesfile)
}
\keyword{datasets}
