% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{An example data set}
\format{An object of class \code{"data.frame"}.}
\usage{
data(ExampleData)
}
\description{
An example data set containing the expression values of the probe sets of two genes.
}
\examples{
data(ExampleData)
}
\keyword{datasets}
