% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleExonLevel}
\alias{ExampleExonLevel}
\title{Exon level data}
\format{An object of class \code{"data.frame"}.}
\usage{
data(ExampleExonLevel)
}
\description{
An example data set containing the summarized expression values on the exon level of the probe sets of one gene.
}
\examples{
data(ExampleExonLevel)
}
\keyword{datasets}
