% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleFirmaOutput}
\alias{ExampleFirmaOutput}
\title{Example data set for FIRMA}
\format{An object of class \code{"data.frame"}.}
\usage{
data(ExampleFirmaOutput)
}
\description{
An example data set for the FIRMA functions
}
\examples{
data(ExampleFirmaOutput)
}
\keyword{datasets}
