% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleGeneLevel}
\alias{ExampleGeneLevel}
\title{Gene level data}
\format{An object of class \code{"data.frame"}.}
\usage{
data(ExampleGeneLevel)
}
\description{
An example data set containing the summarized expression value of one gene.
}
\examples{
data(ExampleGeneLevel)
}
\keyword{datasets}
