% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{FilterInformativeGenes}
\alias{FilterInformativeGenes}
\title{"FilterInformativeGenes"}
\usage{
FilterInformativeGenes(Data)
}
\arguments{
\item{Data}{The output of the REIDS function.}
}
\value{
A subset of the list provided in Data.
}
\description{
The FilterInformativeGenes function works on the list returned by the REIDS functions. It returns a list that only retains the genes that have informative probesets.
}

