% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GetIntensities}
\alias{GetIntensities}
\title{"GetIntensities"}
\usage{
GetIntensities(trans, Data, Groups)
}
\arguments{
\item{trans}{The transcript ID.}

\item{Data}{The data frame with the exon level expression of the transcript.}

\item{Groups}{A list with the groups (columns) of interest}
}
\description{
The GetIntensities function retrieves the exon expression level in the data.
}
