% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{iniREIDS}
\alias{iniREIDS}
\title{"iniREIDS"}
\usage{
iniREIDS(SubgeneData, nsim = 1000)
}
\arguments{
\item{SubgeneData}{A subset of the data. Particularly, a subset of the data corresponding to one gene.}

\item{nsim}{The number of iterations to perform.}
}
\value{
A list with an item per gene. Per gene it is mentioned wich probesets are informative and which are not.
}
\description{
The function is part of the larger REIDS function and performs the I/NI calls filtering model on the data for a single gene to select only the informative probesets. The method is performed if Informative=TRUE in the REIDS function before
before applying the REIDS model itself. The function is written to fit in the flow of the REIDS model
}

