% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{node_coords}
\alias{node_coords}
\title{Node Coordinates}
\usage{
node_coords(edgelist, sorted = FALSE, decreasing = FALSE, ordering = NULL,
  labels = NULL)
}
\arguments{
\item{edgelist}{basically a two-column matrix with edges}

\item{sorted}{logical to indicate if nodes should be sorted}

\item{decreasing}{logical to indicate type of sorting}

\item{ordering}{optional numeric vector providing the ordering of nodes}

\item{labels}{character vector with labels for the nodes}
}
\value{
a vector with the location of nodes in the x-axis
}
\description{
Computes axis locations of each node. This function can be helpful when 
you want to separately plot the node labels using the function mtext.
}
\author{
Gaston Sanchez
}
