% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REIDS.R
\docType{package}
\name{REIDS}
\alias{REIDS}
\alias{REIDS-package}
\title{Random Effects for the Identification of Differential Splicing.}
\description{
The REIDS package contains the REIDS mixed model for the identification of differntially spliced exons.
}

