% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{REIDS_IsoformAssesment}
\alias{REIDS_IsoformAssesment}
\title{REIDS_IsoformAssesment}
\usage{
REIDS_IsoformAssesment(geneIDs, IsoformInfo, ExonLevel, Groups, paired,
  Location = NULL, Name = "REIDSIsoforms")
}
\arguments{
\item{geneIDs}{A vector with the geneIDs to analyze.}

\item{IsoformInfo}{The path to the Composition file created by REIDS_JunctionAssesment or CreateOutput.}

\item{ExonLevel}{The path to the ExonLevel.txt file}

\item{Groups}{A list with  elements speficifing the columns of the data in each group.}

\item{paired}{Logical. Are the groups paired samples?}

\item{Location}{A character string indication the place where the outputs are saved.}

\item{Name}{A character string with the name of the ouput file. Defaults to "REIDSIsoforms".}
}
\value{
The function returns three files. The first file has name "Name_IsoformIndication.txt" and contains an assesment of the relative expression levels of the isoforms. A second file,"Name_ExonTesting.txt", shows information regarding the differential expression of exon. A final file,"Name_PossibelDEIsoforms" lists isoforms which might be differentially expressed between the groups.
}
\description{
The REIDS_IsoformAssesment is an experimental function to analyze the isoform information based on the exon level values and the isoform composition.
}
