% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{REIDSmodel_intern}
\alias{REIDSmodel_intern}
\title{"REIDSmodel_intern"}
\usage{
REIDSmodel_intern(SubgeneData, nsim = 100)
}
\arguments{
\item{SubgeneData}{A subset of the data. Particularly, a subset of the data corresponding to one gene.}

\item{nsim}{The number of iterations to perform.}
}
\value{
A list with 2 items per gene. The first item is the exon scores of the corresponding probesets and the second contains a data frame with the array scores of the exons across the samples.
If the iniREIDS model was performed. The items will be added to the previously made list.
}
\description{
The function is part of the larger REIDS function and performs the REIDS model on the data for a single gene.
}

