% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TC12000010}
\alias{TC12000010}
\title{Example data set}
\format{An object of class \code{"data.frame"}.}
\usage{
data(TC12000010)
}
\description{
An example data set containing the expression values of the probe sets of transcript cluster TC12000010.
}
\examples{
data(TC12000010)
}
\keyword{datasets}
