% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TC12000010_ExonLevel}
\alias{TC12000010_ExonLevel}
\title{Exon level data}
\format{An object of class \code{"data.frame"}.}
\usage{
data(TC12000010_ExonLevel)
}
\description{
An example data set containing the summarized expression values on the exon level of the probe sets of TC12000010.
}
\examples{
data(TC12000010_ExonLevel)
}
\keyword{datasets}
