% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TC12000010_GeneLevel}
\alias{TC12000010_GeneLevel}
\title{Gene level data}
\format{An object of class \code{"data.frame"}.}
\usage{
data(TC12000010_GeneLevel)
}
\description{
An example data set containing the summarized expression value of TC12000010.
}
\examples{
data(TC12000010_GeneLevel)
}
\keyword{datasets}
