% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{TC12000010_positions}
\alias{TC12000010_positions}
\title{Positions}
\format{An object of class \code{"data.frame"}.}
\usage{
data(TC12000010_positions)
}
\description{
Genome location of the probe sets of TC12000010
}
\examples{
data(TC12000010_positions)
}
