% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TC12000010_REIDS_Output}
\alias{TC12000010_REIDS_Output}
\title{REIDS Output}
\format{An object of class \code{"data.frame"}.}
\usage{
data(TC12000010_REIDS_Output)
}
\description{
An example data set containing the obtained REIDS values of TC12000010.
}
\examples{
data(TC12000010_REIDS_Output)
}
\keyword{datasets}
