% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{TC12000010_transcript.clusters}
\alias{TC12000010_transcript.clusters}
\title{Transcript Clusters}
\format{An object of class \code{"data.frame"}.}
\usage{
data(TC12000010_transcript.clusters)
}
\description{
Transcript cluster compositions of TC12000010
}
\examples{
data(TC12000010_transcript.clusters)
}
\keyword{datasets}
