% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TC1500264}
\alias{TC1500264}
\title{Example data set}
\format{An object of class \code{"data.frame"}.}
\usage{
data(TC1500264)
}
\description{
An example data set containing the expression values of the probe sets of transcript cluster TC1500264.
}
\examples{
data(TC1500264)
}
\keyword{datasets}
