% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{trim}
\alias{trim}
\title{"trim"}
\usage{
trim(s)
}
\arguments{
\item{s}{Character vector to trim.}
}
\description{
The trim function trims a long character string on gene annotation to a short single gene annotation.
}
