% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleFirmaOutput}
\alias{ExampleFirmaOutput}
\title{FIRMA Output}
\format{An object of class \code{"data.frame"}.}
\usage{
data(ExampleFirmaOutput)
}
\description{
An example data set containing the obtained FIRMA values for two genes.
}
\examples{
data(ExampleFirmaOutput)
}
\keyword{datasets}
