% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ExonTesting}
\alias{ExonTesting}
\title{"ExonTesting"}
\usage{
ExonTesting(ExonScores, ArrayScores, Exonthreshold = NULL, Groups = list(),
  paired = FALSE, significancelevel = NULL)
}
\arguments{
\item{ExonScores}{The path to the file with the exon scores of the probe sets.}

\item{ArrayScores}{The path to the file with the array scores of the probe sets.}

\item{Exonthreshold}{The exon score threshold to be maintained. If not NULL, probe sets with an exon score lower than this value are not considered further and the p-values will be adjusted for multiplicity after testing. If NULL, all probesets are considered and a multiplicity correction is not performed.}

\item{Groups}{A list with elements specifying the columns of the data in each group.}

\item{paired}{Logical. Are the groups paired? If TRUE the mean paired differences are calculated and tested whether these are significantly different from zero or not.}

\item{significancelevel}{The significance level to be maintained on the p-values. The filtering on the significance is conducted only if an Exonthreshold is specified and the p-value are adjusted for multiplicity.}
}
\value{
A data frame with one line per exon. The columns contain the gene ID, the exon ID, the exon score, the test statistic, a p-value and an adjusted p-value. If the groups are paired also the mean paired difference is given. The p-values are adjusted for multiplicity and filtered on significance if significancelevel is not NULL.
}
\description{
The ExonTesting function performs a t-test (2 groups) or F-test (more than 2 groups) between the array scores of predefined groups. If specified, probe sets are filtered out on exon scores and significance level. The function is the internal function of ASExons.
}
