% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PlotFunction}
\alias{PlotFunction}
\title{"PlotFunction"}
\usage{
PlotFunction(GeneID = NULL, ExonID = NULL, Data, REIDS_Output,
  GeneLevelData = NULL, ExonLevelData = NULL, FIRMA_Data = NULL, groups,
  plottype = "new", location)
}
\arguments{
\item{GeneID}{The gene ID of the gene of interest.}

\item{ExonID}{The exon ID of the exon of interest.}

\item{Data}{The processed data as returned by DataProcessing. This is were the observed probe intensities will be retrieved.}

\item{REIDS_Output}{The output of the REIDS model. This is were the array scores will be retrieved.}

\item{GeneLevelData}{The gene level summarized data to retrieve the gene level values.}

\item{ExonLevelData}{The exon level summarized data to retrieve the exon level values.}

\item{FIRMA_Data}{The output of the FIRMA model to retrieve the FIRMA scores of the samples.}

\item{groups}{The groups of interest in the data.}

\item{plottype}{Should be one of "pdf","new" or "sweave". If "pdf", a location should be provided in "location" and the figure is saved there. If "new" a new graphic device is opened and if "sweave", the figure is made compatible to appear in a sweave or knitr document, i.e. no new device is opened and the plot appears in the current device or document.}

\item{location}{If plottype is "pdf", a location should be provided in "location" and the figure is saved here. the location should include the name of the plot.}
}
\description{
The PlotFunction produces three plots concerning a specific exon and its corresponding gene.
}

