% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{REIDS_JunctionAssesment}
\alias{REIDS_JunctionAssesment}
\title{REIDS_JunctionAssessment}
\usage{
REIDS_JunctionAssesment(ASProbeSets = c(), JAnnotI, EAnnotI, TrAnnotI, Data,
  Groups = list(c(3, 4, 5), c(6, 7, 8)), Low_AllSamples = c(),
  Low_GSamples = c(), Name)
}
\arguments{
\item{ASProbeSets}{The AS probe sets as identified by ASExons.}

\item{JAnnotI}{A table with the line indexing for the juncion associations.}

\item{EAnnotI}{A table with the line indexing for the exon annotations.}

\item{TrAnnotI}{A table with the line indexing for the transcript annotations.}

\item{Data}{The probe level data.}

\item{Groups}{A list with  elements speficifing the columns of the data in each group.}

\item{Low_AllSamples}{A character vector containing the probe sets which are not DABG in all samples.}

\item{Low_GSamples}{A list with a  character vector per group containing the probe sets which are not DABG in that group.}

\item{Name}{A character string with the name of the ouput file.}
}
\value{
The function returns three files. The first file has name "Name.txt" and contains a line per probe set. It shows the reached decision
regarding the probe set (const/AS/not DABG),its linking and exclusion junctions, the fold change, the AS type and its annotated exons. The second
file is a list of all found transcripts for a particular TC ID. The third file indicates whether a specific transcript is present or absent in a group.
}
\description{
The REIDS_JunctionAssessment functions assess identified AS exons based on their 5'end and 3'end and exclusion junction support.
}
